Adding tests to cover using SAML brokers in an organization
Closes #28732
